requirements.txt
tensorflow==2.14.0
keras
numpy==1.26.4
pillow
django
nltk
scikit-learn
requests
djangorestframework
datasets
matplotlib
pandas